% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ed_temp.R
\name{combine_ed_temp}
\alias{combine_ed_temp}
\title{Combine Ed and Temp Function}
\usage{

  combine_ed_temp(file_path_ed = "R:/heatProjections/data/processed/ed_test_data.csv",
  file_path_prism = "R:/heatProjections/data/processed/prism_test_data.csv",
  output_path = "data/processed/combined_test_data.csv")
}
\arguments{
\item{file_path_ed}{The file path to the PRISM data you wish to process.  Defaults to test data on the R server at CDPH.}
}
\description{
This function allows you to ingest a single year of PRISM daily data processed at the zip code scale in Google EARTH.
}
\examples{
get_prism()
}
\keyword{PRISM}
