% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterED.R
\name{filterED}
\alias{filterED}
\title{Filter ED}
\usage{
filterED(bobbCodes = c(55L, 157L, 159L, 244L, 108L, 2L), year = 2005,
  numDiagnosis = 3)
}
\arguments{
\item{year}{Four digit year as it appears in OSPHD filnes 'cdph_ed_rlnYYYY.csv'. Defauls to 2005}

\item{numDiagnosis}{If 1, only the primary diagnosis will be used. Values of 2 or 3 correspond to including secondary and tertiatry diagnoses. Defaults to 3.}

\item{BobbCodes}{This is a collection of text CCS codes to pull out.}
}
\description{
This function allows you to filter a single year of OSHPD Emergency Department Visits for the Clinical Classification System Codes of interest.
}
\examples{
filterED()
}
\keyword{EDvisits}
