% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPRISM.R
\name{getPRISM}
\alias{getPRISM}
\title{Get PRISM Function}
\usage{
getPRISM(year)
}
\arguments{
\item{year}{Four digit year as appears in the PRISM filename 'prism_YYYY.csv'.}
}
\description{
This function allows you to ingest a single year of PRISM daily data processed at the zip code scale in Google EARTH.
}
\examples{
getPRISM()
}
\keyword{PRISM}
