% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPRISM.R
\name{getPRISM}
\alias{getPRISM}
\title{Get PRISM Function}
\usage{

  getPRISM(file_path = "R:/heatProjections/data/processed/prism_test_data.csv")
}
\arguments{
\item{file_path}{The file path to the PRISM data you wish to process.  Defaults to test data on the R server at CDPH.}
}
\description{
This function allows you to ingest a single year of PRISM daily data processed at the zip code scale in Google EARTH.
}
\examples{
getPRISM()
}
\keyword{PRISM}
